+++ core/src/main/java/org/keycloak/jose/jwk/ECPublicJWK.java
@@ -33,13 +33,6 @@ public class ECPublicJWK extends JWK {
     public static final String X = "x";
     public static final String Y = "y";
 
-    @JsonProperty("x5c")
-    private String[] x509CertificateChain;
-
-    private String sha1x509Thumbprint;
-
-    private String sha256x509Thumbprint;
-
     @JsonProperty(CRV)
     private String crv;
 
@@ -71,31 +64,5 @@ public class ECPublicJWK extends JWK {
 
     public void setY(String y) {
         this.y = y;
-    }
-
-    public String[] getX509CertificateChain() {
-        return x509CertificateChain;
-    }
-
-    public void setX509CertificateChain(String[] x509CertificateChain) {
-        this.x509CertificateChain = x509CertificateChain;
-        if (x509CertificateChain != null && x509CertificateChain.length > 0) {
-            try {
-                sha1x509Thumbprint = PemUtils.generateThumbprint(x509CertificateChain, "SHA-1");
-                sha256x509Thumbprint = PemUtils.generateThumbprint(x509CertificateChain, "SHA-256");
-            } catch (NoSuchAlgorithmException e) {
-                throw new RuntimeException(e);
-            }
-        }
-    }
-
-    @JsonProperty("x5t")
-    public String getSha1x509Thumbprint() {
-        return sha1x509Thumbprint;
-    }
-
-    @JsonProperty("x5t#S256")
-    public String getSha256x509Thumbprint() {
-        return sha256x509Thumbprint;
     }
 }
