+++ crypto/default/src/main/java/org/keycloak/crypto/def/BCCertificateUtilsProvider.java
@@ -212,7 +212,6 @@ public class BCCertificateUtilsProvider implements CertificateUtilsProvider {
             }
             return signerBuilder.build(privateKey);
         } catch (Exception e) {
-            e.printStackTrace();
             throw new RuntimeException("Could not create content signer.", e);
         }
     }
